Add safepoints mapping pointers, values, and live registers for instructions (bug 695075 part 7, r=sstangl).

+++ js/src/ion/BitSet.cpp
@@ -60,7 +60,7 @@ BitSet::init()
     size_t sizeRequired = numWords() * sizeof(*bits_);
 
     TempAllocator *alloc = GetIonContext()->temp;
-    bits_ = (unsigned long *)alloc->allocate(sizeRequired);
+    bits_ = (uint32 *)alloc->allocate(sizeRequired);
     if (!bits_)
         return false;
 
@@ -151,7 +151,7 @@ BitSet::fixedPointIntersect(const BitSet *other)
     bool changed = false;
 
     for (unsigned int i = 0; i < numWords(); i++) {
-        unsigned long old = bits_[i];
+        uint32 old = bits_[i];
         bits_[i] &= other->bits_[i];
 
         if (!changed && old != bits_[i])
